azure_ml_api/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
azureml-core>=1.50.0
azureml-inference-server-http>=0.7.0
